Android build working
The project can be built for Android but controls do not work as they should on it.
I'll keep working on the actual features of the game and the PC release and leave Android builds for the near future.
I had to add C++ code to avoid any further issues when packaging.